fix(ilp): commit pending rows after a timeout if no further ingestion (#1731)
