+++ core/src/main/java/io/questdb/cutlass/line/tcp/DefaultLineTcpReceiverConfiguration.java
@@ -113,6 +113,11 @@ public class DefaultLineTcpReceiverConfiguration implements LineTcpReceiverConfi
     }
 
     @Override
+    public long getCommitTimeout() {
+        return 1000;
+    }
+
+    @Override
     public String getAuthDbPath() {
         return null;
     }
