+++ core/src/main/java/io/questdb/cutlass/line/tcp/LineTcpMeasurementScheduler.java
@@ -109,6 +109,7 @@ class LineTcpMeasurementScheduler implements Closeable {
 
         pubSeq = new MPSequence(queueSize);
 
+        long commitInterval = configuration.getCommitTimeout()/2;
         int nWriterThreads = writerWorkerPool.getWorkerCount();
         if (nWriterThreads > 1) {
             FanOut fanOut = new FanOut();
@@ -120,7 +121,7 @@ class LineTcpMeasurementScheduler implements Closeable {
                         queue,
                         subSeq,
                         milliClock,
-                        configuration.getMaintenanceInterval(),
+                        commitInterval,
                         this
                 );
                 writerWorkerPool.assign(i, (Job) lineTcpWriterJob);
@@ -135,7 +136,7 @@ class LineTcpMeasurementScheduler implements Closeable {
                     queue,
                     subSeq,
                     milliClock,
-                    configuration.getMaintenanceInterval(),
+                    commitInterval,
                     this
             );
             writerWorkerPool.assign(0, (Job) lineTcpWriterJob);
